+++ Content.Shared/Charges/Systems/SharedChargesSystem.cs
@@ -34,7 +34,7 @@ public abstract class SharedChargesSystem : EntitySystem
     /// <summary>
     /// Tries to add a number of charges. If it over or underflows it will be clamped, wasting the extra charges.
     /// </summary>
-    public virtual void AddCharges(EntityUid uid, int change, LimitedChargesComponent? comp = null)
+    public virtual void AddCharges(EntityUid uid, float change, LimitedChargesComponent? comp = null)
     {
         if (!Query.Resolve(uid, ref comp, false))
             return;
@@ -84,7 +84,7 @@ public abstract class SharedChargesSystem : EntitySystem
     /// Gets the limited charges component and returns true if the number of charges remaining is less than the specified value.
     /// Will return false if there is no limited charges component.
     /// </summary>
-    public bool HasInsufficientCharges(EntityUid uid, int requiredCharges, LimitedChargesComponent? comp = null)
+    public bool HasInsufficientCharges(EntityUid uid, float requiredCharges, LimitedChargesComponent? comp = null)
     {
         // can't be empty if there are no limited charges
         if (!Resolve(uid, ref comp, false))
@@ -96,7 +96,7 @@ public abstract class SharedChargesSystem : EntitySystem
     /// <summary>
     /// Uses up a specified number of charges. Must check HasInsufficentCharges beforehand to prevent using with insufficient remaining charges.
     /// </summary>
-    public virtual void UseCharges(EntityUid uid, int chargesUsed, LimitedChargesComponent? comp = null)
+    public virtual void UseCharges(EntityUid uid, float chargesUsed, LimitedChargesComponent? comp = null)
     {
         AddCharges(uid, -chargesUsed, comp);
     }
