+++ Content.Shared/RCD/Systems/RCDSystem.cs
@@ -627,14 +627,14 @@ public sealed partial class RCDDoAfterEvent : DoAfterEvent
     public ProtoId<RCDPrototype> StartingProtoId { get; private set; } = default!;
 
     [DataField]
-    public int Cost { get; private set; } = 1;
+    public float Cost { get; private set; } = 1;
 
     [DataField("fx")]
     public NetEntity? Effect { get; private set; } = null;
 
     private RCDDoAfterEvent() { }
 
-    public RCDDoAfterEvent(NetCoordinates location, Direction direction, ProtoId<RCDPrototype> startingProtoId, int cost, NetEntity? effect = null)
+    public RCDDoAfterEvent(NetCoordinates location, Direction direction, ProtoId<RCDPrototype> startingProtoId, float cost, NetEntity? effect = null)
     {
         Location = location;
         Direction = direction;
